REFACTOR: Finished refactoring gravity and collision

+++ DungeonCrafter/core/src/dev/teamcyan/dungeoncrafter/classes/Velocity.java
@@ -1,31 +1,31 @@
 package dev.teamcyan.dungeoncrafter.classes;
 
 public class Velocity {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Velocity (int x, int  y)
+    public Velocity (double x, double  y)
     {
-        this.x = x;
-        this.y = y;
+        this.x=x;
+        this.y=y;
     }
 
-    public void setX(int x)
+    public void setX(double x)
     {
-        this.x = x;
+        this.x=x;
     }
 
-    public void setY(int  y)
+    public void setY(double  y)
     {
-        this.y = y;
+        this.y=y;
     }
 
-    public int  getY()
+    public double getY()
     {
         return this.y;
     }
 
-    public int getX()
+    public double getX()
     {
         return this.x;
     }
